+++ src/point/new_point.c
@@ -6,13 +6,13 @@
 /*   By: jaberkro <jaberkro@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/24 10:48:21 by jaberkro      #+#    #+#                 */
-/*   Updated: 2022/11/18 14:33:32 by jaberkro      ########   odam.nl         */
+/*   Updated: 2022/12/08 14:57:48 by bsomers       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "point.h"
 
-t_point	new_point(float x, float y, float z)
+t_point	new_point(double x, double y, double z)
 {
 	t_point	point;
